bench 9099336

+++ src/nnue.h
@@ -31,8 +31,8 @@ struct Weights {
 struct Net {
     std::array<int16_t , L1_SIZE * INPUT_SIZE> weights0{};
     std::array<int16_t, L1_SIZE> bias0{};
-    std::array<int16_t, L1_SIZE * L2_SIZE * 2> weights1{};
-    std::array<int16_t, L2_SIZE> bias1{};
+    std::array<float, L1_SIZE * L2_SIZE * 2> weights1{};
+    std::array<float, L2_SIZE> bias1{};
     std::array<float, L2_SIZE * OUTPUT_SIZE> weights2{};
     std::array<float, OUTPUT_SIZE> bias2{};
     std::array<std::array<int16_t, L1_SIZE>, 2> accumulator{};
